+++ tsconfig.app.json
@@ -29,6 +29,7 @@
     "noUncheckedSideEffectImports": true
   },
   "include": [
-    "src"
+    "src",
+    "node_modules/vitest/globals.d.ts"
   ]
 }
